refactor: turn wave-7 Slice C densify four fat non-support suites (#4342)
Extract shared fixtures into sibling *_cases.dart peers for diplomacy
phase handler, army-move trace events, research extraction integration,
and world-market treasury clamp suites. Pin Slice C ≤250-line targets
in turn_resolver_split_wave7_test.

Refs #4342
